# swagger.yml
openapi: 3.0.0
info:
  title: Authentication API
  version: 1.0.0
paths:
  /login:
    get:
      summary: Login
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
            format: uuid      
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /validate:
    post:
      summary: Validate token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequest'
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /refresh:
    post:
      summary: Refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New tokens generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
    LoginResponse:
      type: object
      properties:
        access:
          type: string
          description: JWT access token
        refresh:
          type: string
          description: Base64 encoded refresh token
    ValidateRequest:
      type: object
      properties:
        access: 
          type: string
          description: JWT access token
    ValidateResponse:
      type: object
      properties:
        id:
          type: string
          description: ID extracted from token claims
        status:
          type: string
          description: "valid"
    RefreshRequest:
      type: object
      properties:
        access:
          type: string
          description: JWT access token
        refresh:
          type: string
          description: Base64 encoded refresh token